% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlist_create.R
\name{charlist_create}
\alias{charlist_create}
\title{charlist_creat}
\usage{
charlist_create(dataframe_column)
}
\arguments{
\item{column}{vector}
}
\value{
character list
}
\description{
charlist_creat: removes specified column from dataframe or list, empty if
neither dataframe or list
}
