% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertChar.R
\name{ConvertChar}
\alias{ConvertChar}
\title{ConvertChar
Converts character to numeric string, can handle ordinal varialbes
will convert them to independent binary response variables, which is
appropriate for descriptive analysis, but for other analyses variables
should be kept in their ordinal form. Also determines if any columns contain
dates and removes them. Otherwise would turn each individual date into a separate
categorical variables}
\usage{
ConvertChar(dataframe, group_name)
}
\arguments{
\item{dataframe}{dataframe to be inputted}

\item{group_name}{variable specifying groups}
}
\value{
a dataframe with all columns containing characters, converted to numeric
}
\description{
ConvertChar
Converts character to numeric string, can handle ordinal varialbes
will convert them to independent binary response variables, which is
appropriate for descriptive analysis, but for other analyses variables
should be kept in their ordinal form. Also determines if any columns contain
dates and removes them. Otherwise would turn each individual date into a separate
categorical variables
}
