% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedetect.R
\name{datedetect}
\alias{datedetect}
\title{datedetect
determines if column contains dates. These will be removed}
\usage{
datedetect(dataframe_column)
}
\arguments{
\item{dataframe_column}{A column vector}
}
\value{
boolean
}
\description{
datedetect
determines if column contains dates. These will be removed
}
