% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tab_addnew.R
\name{flex_tab_addnew}
\alias{flex_tab_addnew}
\title{flex-table to output to .docx
takes table (cattable) and ouputs it is a formatted table
in word to specified output file "outfile." Output file must be
format "filename.docx"}
\usage{
flex_tab_addnew(cattable, outfile)
}
\arguments{
\item{cattable}{table to be converted to flextable}

\item{outfile}{file to export table to}
}
\value{
flextable
}
\description{
flex-table to output to .docx
takes table (cattable) and ouputs it is a formatted table
in word to specified output file "outfile." Output file must be
format "filename.docx"
}
