% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_identical.R
\name{is_identical}
\alias{is_identical}
\title{is_identical
determines if all values in column are identical}
\usage{
is_identical(dataframe_column)
}
\arguments{
\item{dataframe_column}{A column vector}
}
\value{
boolean
}
\description{
is_identical
determines if all values in column are identical
}
