% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list_element.R
\name{is_list_element}
\alias{is_list_element}
\title{is_list_element
Determines if a variable is present in a particular list}
\usage{
is_list_element(list, var)
}
\arguments{
\item{variable}{searching for}
}
\value{
boolean
}
\description{
is_list_element
Determines if a variable is present in a particular list
}
