% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_normal.R
\name{is_normal}
\alias{is_normal}
\title{is_normal}
\usage{
is_normal(dataframe_column)
}
\arguments{
\item{dataframe_column}{A column vector}
}
\value{
boolean
}
\description{
Takes a continuous variable from a dataframe
and determines if the values are normally distributed
outputs TRUE if they are, FALSE if not
}
