% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_from_user.R
\name{list_from_user}
\alias{list_from_user}
\title{list_from_user
allows user to input names for the column names
must ensure that there is no text including commented text beneath the final function
call or will not work
input is the number of groups being analyzed}
\usage{
list_from_user(number_of_groups)
}
\arguments{
\item{number}{of groups}
}
\value{
A list of desired column names
}
\description{
list_from_user
allows user to input names for the column names
must ensure that there is no text including commented text beneath the final function
call or will not work
input is the number of groups being analyzed
}
