% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_round.R
\name{p_value_round}
\alias{p_value_round}
\title{p_value_round
inputs a pvalue from a statistical test,
rounds it to two decimal places if > 0.1,
three decmial places if < 0.1
prints < 0.001 if p is < 0.001
returns rounded p value}
\usage{
p_value_round(pval)
}
\arguments{
\item{pval}{A p-value}
}
\value{
A rounded p-value
}
\description{
p_value_round
inputs a pvalue from a statistical test,
rounds it to two decimal places if > 0.1,
three decmial places if < 0.1
prints < 0.001 if p is < 0.001
returns rounded p value
}
