% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsd1.R
\name{statsd1}
\alias{statsd1}
\title{statsd1
Computes table of descriptive statistics and table of p-values from pairwise comparisons}
\usage{
statsd1(dataframe, set_fisher = 10, group_name, number_of_groups)
}
\arguments{
\item{dataframe}{A dataframe}

\item{set_fisher}{Set the value beneath which fisher's exact will be used}

\item{group_name}{A column variable that specifies groups}

\item{number_of_groups}{The number of groups being compared}
}
\value{
A list of tables, one summarizing descriptive statistics, the other a list of p-values from pairwise testing if number of groups > 2
}
\description{
Set fisher allows user to specify the threshold number of observations beneath which fisher's exact test
will be used to compare categorical variables. Above this threshold chi squared will be used.
fxn will determine if continuous variables are normally distributed or not and use t.test or
man-whitney for 2 groups, anova or kriskal wallis for 3 or more groups
}
