% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_var.R
\name{remove_var}
\alias{remove_var}
\title{remove var
remove_var: removes specified variable from list}
\usage{
remove_var(dataframe, var)
}
\arguments{
\item{dataframe}{A column vector}

\item{var}{A variable}
}
\value{
dataframe
}
\description{
remove var
remove_var: removes specified variable from list
}
